:sparkles: 메시지 공용 도메인 서비스 코드 적용
